Throw an error if registry deploy fails
